third_party/WebKit/LayoutTests/fast/events/wheel/mouse-wheel-scroll-latching.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
  ::-webkit-scrollbar {
    display: none;
  }
  body {
    margin: 0px;
    height: 100px;
    width: 100px;
  }
  #parentDiv {
    background-color: #FF7F7F;
    height: 60px;
    width: 60px;
    overflow: scroll;
  }
  #content1 {
    height: 70px;
    width: 70px;
  }
  #childDiv {
    background-color: #84BE6A;
    height: 50px;
    width: 50px;
    overflow: scroll;
  }
  #content2 {
    height: 60px;
    width: 60px;
  }
</style>

<div id="parentDiv">
  <div id="content1">
    <div id="childDiv">
      <div id="content2">
      </div>
    </div>
  </div>
</div>

<script>
var parentDiv = document.getElementById('parentDiv');
var childDiv = document.getElementById('childDiv');

const WHEEL_SOURCE_TYPE = 2; // MOUSE_INPUT from synthetic_gesture_params.h
function scrollDown(pixels_to_scroll, start_x, start_y, speed_in_pixels_s) {
  return new Promise((resolve, reject) => {
    chrome.gpuBenchmarking.smoothScrollBy(pixels_to_scroll,
                                          resolve,
                                          start_x,
                                          start_y,
                                          WHEEL_SOURCE_TYPE,
                                          'down',
                                          speed_in_pixels_s);
  });
}

const MAX_RAF = 1000;
var last_child_scroll_offset = childDiv.scrollTop;
var last_parent_scroll_offset = parentDiv.scrollTop;
var last_changed_count = 0;
function waitForAnimationEnd() {
  return new Promise((resolve, reject) => {
    function tick(raf_count) {
    // We requestAnimationFrame either for 1000 frames or until 20 frames with
    // no change have been observed.
      if (raf_count >= MAX_RAF || raf_count - last_changed_count > 20) {
        resolve();
      } else {
        if (childDiv.scrollTop != last_child_scroll_offset ||
            parentDiv.scrollTop != last_parent_scroll_offset) {
          last_changed_count = raf_count;
          last_child_scroll_offset = childDiv.scrollTop;
          last_parent_scroll_offset = parentDiv.scrollTop;
        }
        requestAnimationFrame(tick.bind(this, raf_count + 1));
      }
    }
    tick(0);
  });
}

var rect = childDiv.getBoundingClientRect();
function testWheelScrollLatching() {
  return scrollDown(400, (rect.left + rect.right) / 2, (rect.top + rect.bottom) / 2, 1000)
  .then(waitForAnimationEnd)
  .then(() => {
    assert_equals(childDiv.scrollTop, 10, "childDiv must be fully scrolled");
    assert_equals(parentDiv.scrollTop, 0, "parentDiv shouldn't scroll at all");
  });
}

promise_test(t => {
    return testWheelScrollLatching();
  }, 'With wheel scroll latching enabled only child div must scroll');
</script>
